Remove `idempotent` property of encoding
The idea was that, if we knew that encode() and decode() were
idempotent, we could in certain situations skip a decode step in
abstract-leveldown and avoid allocating a callback function. For
example if leveldown returned a string to be decoded with the utf8
encoding; that would be a noop we could skip.

It worked fine. But I didn't get to test that it had a significant
performance benefit. And to avoid painting ourselves into a corner
with this optimization (for example, if we need to normalize data in
decode() like level-js does) I'm not gonna spend more time on it.